Use common 'Pinch' gesture instead of Distance.
Pinch reports a 'change' piece of data that allows the user to
distinguish between what were previously termed Pinch and Expand.
